+++ LiteCore/Support/EncryptedStream.cc
@@ -108,7 +108,7 @@ namespace litecore {
                                  plaintext);
         _output->write(ciphertext);
         LogTo(BlobLog, "WRITE #%2llu: %llu bytes, final=%d --> %llu bytes ciphertext",
-            _blockID-1, (uint64_t)plaintext.size, finalBlock, (uint64_t)ciphertext.size);
+            (unsigned long long)(_blockID-1), (unsigned long long)plaintext.size, finalBlock, (unsigned long long)ciphertext.size);
 #else
         error::_throw(error::Unimplemented);
 #endif
@@ -197,7 +197,7 @@ namespace litecore {
                       finalBlock,
                       output, slice(blockBuf, bytesRead));
         LogTo(BlobLog, "READ  #%2llu: %llu bytes, final=%d --> %llu bytes ciphertext",
-            _blockID-1, (uint64_t)bytesRead, finalBlock, (uint64_t)outputSize);
+            (unsigned long long)(_blockID-1), (unsigned long long)bytesRead, finalBlock, (unsigned long long)outputSize);
         return outputSize;
 #else
         error::_throw(error::Unimplemented);
@@ -264,7 +264,7 @@ namespace litecore {
         uint64_t blockID = min(pos / kFileBlockSize, _finalBlockID);
         uint64_t blockPos = blockID * kFileBlockSize;
         if (blockID != _bufferBlockID) {
-            LogTo(BlobLog, "SEEK %llu (block %llu + %llu bytes)", pos, blockID, pos - blockPos);
+            LogTo(BlobLog, "SEEK %llu (block %llu + %llu bytes)", (unsigned long long)pos, (unsigned long long)blockID, (unsigned long long)(pos - blockPos));
             _input->seek(blockPos);
             _blockID = blockID;
             fillBuffer();
